requirements.txt
pandas
pandas_ta
websockets
python-dotenv
aiohttp
